+++ ChangeLog.md
@@ -3,3 +3,15 @@
 v0.1
 
 - Core files
+
+v0.2
+
+- Added tests
+- Added entities
+- Added deps
+- Added controller
+
+v0.3
+
+- Added create character route
+- Added persisting character